[rb] remove support for legacy Edge HTML browser
