+++ rb/lib/selenium/webdriver.rb
@@ -39,8 +39,6 @@ module Selenium
     autoload :Chrome,     'selenium/webdriver/chrome'
     autoload :DevTools,   'selenium/webdriver/devtools'
     autoload :Edge,       'selenium/webdriver/edge'
-    autoload :EdgeHtml,   'selenium/webdriver/edge'
-    autoload :EdgeChrome, 'selenium/webdriver/edge'
     autoload :Firefox,    'selenium/webdriver/firefox'
     autoload :IE,         'selenium/webdriver/ie'
     autoload :Remote,     'selenium/webdriver/remote'
