+++ rb/lib/selenium/webdriver/common/service.rb
@@ -45,14 +45,6 @@ module Selenium
           Edge::Service.new(**opts)
         end
 
-        def edge_chrome(**opts)
-          EdgeChrome::Service.new(**opts)
-        end
-
-        def edge_html(**opts)
-          EdgeHtml::Service.new(**opts)
-        end
-
         def safari(**opts)
           Safari::Service.new(**opts)
         end
