+++ rb/lib/selenium/webdriver/edge.rb
@@ -21,32 +21,12 @@ require 'net/http'
 
 module Selenium
   module WebDriver
-    module EdgeHtml
-      autoload :Driver, 'selenium/webdriver/edge_html/driver'
-      autoload :Options, 'selenium/webdriver/edge_html/options'
-      autoload :Service, 'selenium/webdriver/edge_html/service'
-
-      def self.driver_path=(path)
-        WebDriver.logger.deprecate 'Selenium::WebDriver::Edge#driver_path=',
-                                   'Selenium::WebDriver::Edge::Service#driver_path=',
-                                   id: :driver_path
-        Selenium::WebDriver::Edge::Service.driver_path = path
-      end
-
-      def self.driver_path
-        WebDriver.logger.deprecate 'Selenium::WebDriver::Edge#driver_path',
-                                   'Selenium::WebDriver::Edge::Service#driver_path',
-                                   id: :driver_path
-        Selenium::WebDriver::Edge::Service.driver_path
-      end
-    end # EdgeHtml
-
-    module EdgeChrome
-      autoload :Bridge, 'selenium/webdriver/edge_chrome/bridge'
-      autoload :Driver, 'selenium/webdriver/edge_chrome/driver'
-      autoload :Profile, 'selenium/webdriver/edge_chrome/profile'
-      autoload :Options, 'selenium/webdriver/edge_chrome/options'
-      autoload :Service, 'selenium/webdriver/edge_chrome/service'
+    module Edge
+      autoload :Bridge, 'selenium/webdriver/edge/bridge'
+      autoload :Driver, 'selenium/webdriver/edge/driver'
+      autoload :Profile, 'selenium/webdriver/edge/profile'
+      autoload :Options, 'selenium/webdriver/edge/options'
+      autoload :Service, 'selenium/webdriver/edge/service'
 
       def self.path=(path)
         Platform.assert_executable path
@@ -56,8 +36,6 @@ module Selenium
       def self.path
         @path ||= nil
       end
-    end # EdgeHtml
-
-    Edge = EdgeChrome
+    end # Edge
   end # WebDriver
 end # Selenium
