+++ rb/lib/selenium/webdriver/remote/capabilities.rb
@@ -82,17 +82,6 @@ module Selenium
           end
 
           def edge(opts = {})
-            edge_chrome(opts)
-          end
-
-          def edge_html(opts = {})
-            new({
-              browser_name: 'MicrosoftEdge',
-              platform_name: :windows
-            }.merge(opts))
-          end
-
-          def edge_chrome(opts = {})
             new({
               browser_name: 'MicrosoftEdge'
             }.merge(opts))
