+++ rb/spec/unit/selenium/webdriver/edge/driver_spec.rb
@@ -21,19 +21,18 @@ require File.expand_path('../spec_helper', __dir__)
 
 module Selenium
   module WebDriver
-    module EdgeHtml
+    module Edge
       describe Driver do
         let(:service) { instance_double(Service, launch: service_manager) }
         let(:service_manager) { instance_double(ServiceManager, uri: 'http://example.com') }
         let(:valid_response) do
           {status: 200,
-           body: {value: {sessionId: 0, capabilities: Remote::Capabilities.edge_html}}.to_json,
+           body: {value: {sessionId: 0, capabilities: Remote::Capabilities.edge}}.to_json,
            headers: {"content_type": "application/json"}}
         end
 
         def expect_request(body: nil, endpoint: nil)
-          body = (body || {capabilities: {firstMatch: [browserName: "MicrosoftEdge",
-                                                       platformName: "windows"]}}).to_json
+          body = (body || {capabilities: {firstMatch: [browserName: "MicrosoftEdge"]}}).to_json
           endpoint ||= "#{service_manager.uri}/session"
           stub_request(:post, endpoint).with(body: body).to_return(valid_response)
         end
@@ -50,18 +49,16 @@ module Selenium
 
         context 'with :desired capabilities' do
           it 'accepts value as a Symbol' do
-            expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge",
-                                                              platformName: 'windows']}})
+            expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge"]}})
+
             expect {
-              expect { Driver.new(desired_capabilities: :edge_html) }.to have_deprecated(:desired_capabilities)
+              expect { Driver.new(desired_capabilities: :edge) }.to have_deprecated(:desired_capabilities)
             }.not_to raise_exception
           end
 
-          it 'accepts Capabilities.edge_html' do
-            capabilities = Remote::Capabilities.edge_html(invalid: 'foobar')
-            expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge",
-                                                              platformName: 'windows',
-                                                              invalid: 'foobar']}})
+          it 'accepts Capabilities.edge' do
+            capabilities = Remote::Capabilities.edge(invalid: 'foobar')
+            expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge", invalid: 'foobar']}})
 
             expect {
               expect { Driver.new(desired_capabilities: capabilities) }.to have_deprecated(:desired_capabilities)
@@ -70,8 +67,7 @@ module Selenium
 
           it 'accepts constructed Capabilities with Snake Case as Symbols' do
             capabilities = Remote::Capabilities.new(browser_name: 'MicrosoftEdge', invalid: 'foobar')
-            expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge",
-                                                              invalid: 'foobar']}})
+            expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge", invalid: 'foobar']}})
 
             expect {
               expect { Driver.new(desired_capabilities: capabilities) }.to have_deprecated(:desired_capabilities)
@@ -80,8 +76,7 @@ module Selenium
 
           it 'accepts constructed Capabilities with Camel Case as Symbols' do
             capabilities = Remote::Capabilities.new(browserName: 'MicrosoftEdge', invalid: 'foobar')
-            expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge",
-                                                              invalid: 'foobar']}})
+            expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge", invalid: 'foobar']}})
 
             expect {
               expect { Driver.new(desired_capabilities: capabilities) }.to have_deprecated(:desired_capabilities)
@@ -117,9 +112,8 @@ module Selenium
         end
 
         it 'accepts provided Options as sole parameter' do
-          opts = {start_page: 'http://selenium.dev'}
-          expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge",
-                                                            "ms:startPage": "http://selenium.dev"]}})
+          opts = {invalid: 'foobar', args: ['-f']}
+          expect_request(body: {capabilities: {firstMatch: ["browserName": "MicrosoftEdge", "ms:edgeOptions": opts]}})
 
           expect {
             expect { Driver.new(options: Options.new(**opts)) }.to have_deprecated(:browser_options)
@@ -127,12 +121,11 @@ module Selenium
         end
 
         it 'accepts combination of Options and Capabilities' do
-          caps = Remote::Capabilities.edge_html(invalid: 'foobar')
-          browser_opts = {start_page: 'http://selenium.dev'}
-          expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge",
-                                                            platformName: "windows",
-                                                            invalid: "foobar",
-                                                            "ms:startPage": "http://selenium.dev"]}})
+          caps = Remote::Capabilities.edge(invalid: 'foobar')
+          browser_opts = {args: ['-f']}
+          expect_request(body: {capabilities: {firstMatch: ["browserName": "MicrosoftEdge",
+                                                            "invalid": "foobar",
+                                                            "ms:edgeOptions": browser_opts]}})
 
           expect {
             expect {
@@ -148,16 +141,13 @@ module Selenium
 
         context 'with :capabilities' do
           it 'accepts value as a Symbol' do
-            expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge",
-                                                              platformName: 'windows']}})
-            expect { Driver.new(capabilities: :edge_html) }.not_to raise_exception
+            expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge"]}})
+            expect { Driver.new(capabilities: :edge) }.not_to raise_exception
           end
 
-          it 'accepts Capabilities.edge_html' do
-            capabilities = Remote::Capabilities.edge_html(invalid: 'foobar')
-            expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge",
-                                                              platformName: 'windows',
-                                                              invalid: 'foobar']}})
+          it 'accepts Capabilities.edge' do
+            capabilities = Remote::Capabilities.edge(invalid: 'foobar')
+            expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge", invalid: 'foobar']}})
 
             expect { Driver.new(capabilities: capabilities) }.not_to raise_exception
           end
@@ -211,11 +201,22 @@ module Selenium
             end
 
             it 'with Options instance' do
-              browser_opts = {start_page: 'http://selenium.dev'}
+              options = Options.new(args: ['-f'])
               expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge",
-                                                                'ms:startPage': 'http://selenium.dev']}})
+                                                                'ms:edgeOptions': {'args': ['-f']}]}})
 
-              expect { Driver.new(capabilities: [Options.new(**browser_opts)]) }.not_to raise_exception
+              expect { Driver.new(capabilities: [options]) }.not_to raise_exception
+            end
+
+            it 'with Options instance with profile' do
+              profile = Profile.new.tap(&:layout_on_disk)
+              allow(profile).to receive(:directory).and_return("PROF_DIR")
+              options = Options.new(profile: profile)
+              expect_request(body: {capabilities:
+                                      {firstMatch: [browserName: "MicrosoftEdge",
+                                                    'ms:edgeOptions': {"args": ["--user-data-dir=PROF_DIR"]}]}})
+
+              expect { Driver.new(capabilities: [options]) }.not_to raise_exception
             end
 
             it 'with Capabilities instance' do
@@ -227,17 +228,16 @@ module Selenium
 
             it 'with Options instance and an instance of a custom object responding to #as_json' do
               expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge",
-                                                                # 'ms:startPage': 'http://selenium.dev',
+                                                                'ms:edgeOptions': {},
                                                                 'company:key': 'value']}})
               expect { Driver.new(capabilities: [Options.new, as_json_object.new]) }.not_to raise_exception
             end
 
             it 'with Options instance, Capabilities instance and instance of a custom object responding to #as_json' do
               capabilities = Remote::Capabilities.new(browser_name: 'MicrosoftEdge', invalid: 'foobar')
-              options = Options.new(start_page: 'http://selenium.dev')
-              expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge",
-                                                                invalid: 'foobar',
-                                                                'ms:startPage': 'http://selenium.dev',
+              options = Options.new(args: ['-f'])
+              expect_request(body: {capabilities: {firstMatch: [browserName: "MicrosoftEdge", invalid: 'foobar',
+                                                                'ms:edgeOptions': {'args': ['-f']},
                                                                 'company:key': 'value']}})
 
               expect { Driver.new(capabilities: [capabilities, options, as_json_object.new]) }.not_to raise_exception
@@ -245,6 +245,6 @@ module Selenium
           end
         end
       end
-    end # EdgeHtml
+    end # Chrome
   end # WebDriver
 end # Selenium
