+++ rb/spec/unit/selenium/webdriver/edge/profile_spec.rb
@@ -21,7 +21,7 @@ require File.expand_path('../spec_helper', __dir__)
 
 module Selenium
   module WebDriver
-    module EdgeChrome
+    module Edge
       describe Profile do
         let(:profile) { Profile.new }
         let(:model) { '/some/path' }
@@ -67,6 +67,6 @@ module Selenium
           expect(result['some']['other']['pref']).to eq(123)
         end
       end
-    end # EdgeChrome
+    end # Edge
   end # WebDriver
 end # Selenium
